Fix bug for cellcyle PCA plots and other minor fixes
The PCA plots to visualise cellcycle effects have not been based
on the given cellcycle genes since the switch to Seurat v3 due to
the change of a parameter name in RunPCA from "pc.genes" to "features".
The parameter name is now updated. Please note that this means that
any cell cycle PCA plots made by the pipeline with Seurat v3 will
have been based on variable genes (which are the default features).

Also:
- Fix colors for violin plots of known genes with Leiden clustering
- Fix export of diffusion map coordinates
- Make geneset analysis optional