Updates and refactoring to the support analysis of large datasets
(1) The beginSeurat task is now split into:

* createSeuratObject: for creating, QC'ing and subsetting an initial seurat object.
* exploreHvgAndCellCycle: for making plots of HVG and visualising cell cycle variation
* normaliseAndScale: perform normalisation and scaling
* seuratPCA: perform PCA

(2) Support for use of cc.genes has been added. If symbols are not supplied by the
user, these lists will be used and automatically translated from human to mouse.
User supplied lists must now comprise of a single column of gene symbols.

(3) Cluster statistics are now computed in seperate, per-cluster tasks rather
than when summarising the markers.

(4) Job threads are specified with a help function "resources.get" which will compute
the resource requirements for slurm and return a dictionary that can be used to
update e.g. locals().

(5) Make more tasks optional. All of the optional tasks are now controlled
in a single place at the beginning of the pipeline.yml file:

run:
  explore_hvg_and_cell_cycle: True
  singleR: False
  jackstraw: False
  compare_clusters: True
  characterise_markers: True
  diffusionmap: False
  phate: True
  paga: True
  velocity: False
  knownmarkers: False
  exprsreport: False
  genesets: False
  cellbrowser: False

(6) PDF plots are now disabled by default for many tasks (WIP). They can
be generated by setting the plot_pdf option to "True". But in general
it is recommended to make them on demand by rerunning individual R scripts
as needed in order to preserve disk space.