Switch to scanpy for knn graph, cluster and umap computation.
This is done mainly for performance reasons.

For knn scanpy is faster than use of RANN in Seurat because
it uses an approximate method. Scany uses pynndescent, which
outperforms annoy, the approximate method that is wrapped
by Seurat, see: https://erikbern.com/2018/06/17/new-approximate-nearest-neighbor-benchmarks.html.
In fact, it would be (much) better to be using HNSW - this has been
implemented already in pegasus (see: https://pegasus.readthedocs.io/en/0.x/
and https://doi.org/10.1101/823682) but it is not so straightforward
to install.

Use of Leiden clustering with Seurat for large datasets is
currently not possible so use of scanpy is now essential.

In practise the pipeline now returns extremely similar results to
those seen with the pure Seurat workflow, in far less time.

Additionally:

- replace use of "txt" with "tsv" throughout the repo for
  consistency.

- make it possible to link in results for certain tasks
  from previous runs (see the "headstart" section in pipeline.yml).

- clusters with very small numbers of cells (<10) and now
  merged into a catch all "911" cluster so that they do not
  dominate plots. This may not be appropriate behaviour
  for small datasets!